requirements.txt
asyncpg
aiohttp
requests
more-itertools
sqlalchemy
python-dotenv